Update README.md
# School Data Analysis Project

## Overview of Project
The project involves cleaning student data, which comprises misspellings, changes of the case, duplicated names, and special characters that are not supposed to be in the data.

### Purpose
This project aims to analyze students' data from different school districts and clean the data so it's free from errors. The student's scores were a focal point and needed to be cleaned so the code can run properly.

## Analysis and Challenges
The data had a lot of prefixes added to the student's name, for example, Dr. A student in high school doesn't have a doctorate so that's an error. The data was cleaned from all errors and the scores that indicated zero was replaced by NAN. It is necessary to replace the zeros so it doesn't affect the entire data when we calculating the average or percentage of student grades. The data are given to users to analyze the passing grade from each district by comparing their maths in reading. the change in the grade finally gave an overview of the schools did averagely on maths and reding from 9th grade to 12th grade.

##Result

Summary.
The project was focused on cleaning student data that enabled clarity. There was a need for data replacement in the Thomas High School data, the data indicating zero was replaced by a NAN to enable accurate calculations. Replacing the zero score to nan actually modify the data to omit further calculation errors but it also dropped Thomas High School general district grading to a lower position which was expected.