Update README.md
# School Data Analysis Project

## Overview of Project
The project involves cleaning student data which comprises misspellings, changes of the case, duplicated names, and special characters that are not supposed to be in the data.

### Purpose
The purpose of this project is to analyze students' data from the different school districts, clean the data so it is free from errors. The student's scores were a focal point and needed to be cleaned so the code can run properly.

## Analysis and Challenges
The data had a lot of prefixes added to the student's name, for example, Dr. A student in high school doesn't have a doctorate so that's an error. The data was cleaned from all errors and the scores that indicated zero was replaced by NAN. It is necessary to replace the zeros so it doesn't affect the entire data when we calculating the average or percentage of student grades. The data are given to be used to analyze the passing grade from each district by comparing their maths in reading. the change in the grade finally gave an overview of the schools did averagely on maths and reding from 9th grade to 12th grade.

## Results
(https://user-images.githubusercontent.com/86568537/130340099-dcb223d4-84d6-4e94-bf8e-2bdbe40f0372.png)
Summary
Calculating for both math and reading average scores and percentage would have been impossible if the data was not been properly cleaned. Cleaning the data eliminates the prefix errors being attached to high school student names. Most data was replaced for ease of calculation. The zeros data in Thomas High School was replaced by a nan so it will be easy to calculate for average grades. Although the replaced data dropped Thomas High school rating but the data is now cleared from errors.